Renamed Mesh to MeshGen to avoid conflicts
